# specs/002-read-the-specification/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GhostCart AP2 Demo API
  description: |
    REST API for GhostCart demonstrating Agent Payments Protocol v0.1 with AWS Strands SDK.

    This API provides endpoints for:
    - Chat interaction with AI shopping agents (Supervisor, HP Shopping, HNP Delegate)
    - Server-Sent Events for real-time agent streaming
    - Mandate signing for user authorization
    - Transaction chain retrieval for audit trails
    - Monitoring job management for autonomous purchasing
    - Product search and payment method retrieval

    Protocol Compliance: All mandate structures match AP2 v0.1 specification from https://ap2-protocol.org/specification/
  version: 1.0.0
  contact:
    name: GhostCart Team
    email: support@ghostcart-demo.com
servers:
  - url: http://localhost:8000/api
    description: Local development server
...
paths:
...
  /monitoring/jobs:
    get:
      tags: [Monitoring]
      summary: List monitoring jobs for user
      description: Retrieves all monitoring jobs (active, completed, expired) for specified user.
      operationId: listMonitoringJobs
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: User identifier
          example: "user_demo_001"
      responses:
        '200':
          description: Monitoring jobs list
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/MonitoringJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /monitoring/jobs/{job_id}:
    delete:
      tags: [Monitoring]
      summary: Cancel monitoring job
      description: Cancels active monitoring job and removes from APScheduler.
      operationId: cancelMonitoringJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Job identifier
          example: "job_mon_xyz789"
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    ConstraintsObject:
      type: object
      required: [max_price_cents, max_delivery_days, currency]
      properties:
        max_price_cents:
          type: integer
          minimum: 1
          description: Maximum price in cents
          example: 18000
        max_delivery_days:
          type: integer
          minimum: 1
          maximum: 30
          description: Maximum delivery time in days
          example: 2
        currency:
          type: string
          enum: [USD]
          example: "USD"
...
    MonitoringJob:
      type: object
      required: [job_id, intent_mandate_id, user_id, product_query, constraints, schedule_interval_minutes, active, created_at, expires_at]
      properties:
        job_id:
          type: string
          example: "job_mon_xyz789"
        intent_mandate_id:
          type: string
          example: "intent_hnp_abc123"
        user_id:
          type: string
          example: "user_demo_001"
        product_query:
          type: string
          example: "Apple AirPods Pro"
        constraints:
          $ref: '#/components/schemas/ConstraintsObject'
        schedule_interval_minutes:
          type: integer
          minimum: 1
          example: 5
        active:
          type: boolean
          example: true
        last_check_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-10-17T14:30:00Z"
        created_at:
          type: string
          format: date-time
          example: "2025-10-17T14:00:00Z"
        expires_at:
          type: string
          format: date-time
          example: "2025-10-24T14:00:00Z"
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: AP2 standard error code
          enum:
            - ap2:mandate:chain_invalid
            - ap2:mandate:signature_invalid
            - ap2:mandate:expired
            - ap2:mandate:constraints_violated
            - ap2:credentials:unavailable
            - ap2:payment:declined
            - api:bad_request
            - api:not_found
            - api:internal_error
          example: "ap2:mandate:signature_invalid"
        message:
          type: string
          description: Human-readable error description
          example: "Cart mandate signature verification failed"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "api:bad_request"
            message: "Missing required field: session_id"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "api:not_found"
            message: "Transaction not found: txn_invalid"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "api:internal_error"
            message: "An unexpected error occurred"